feat(frontend): render media load hints with lazy session detail
