+++ packages/i18n/messages/ar.json
@@ -1,5 +1,25 @@
 {
 	"$schema": "https://inlang.com/schema/inlang-message-format",
+	"settings.mediaLoad.title": "حمل الوسائط",
+	"settings.mediaLoad.description": "قراءات برنامج التشغيل لكل نواة والجلسات المرتبطة",
+	"settings.mediaLoad.details": "تفاصيل الوسائط",
+	"settings.mediaLoad.coreCount": "الأنوية: {count}",
+	"settings.mediaLoad.encode": "الترميز",
+	"settings.mediaLoad.decode": "فك الترميز",
+	"settings.mediaLoad.jpeg": "فك ترميز JPEG",
+	"settings.mediaLoad.rga": "RGA · معالجة الصور",
+	"settings.mediaLoad.load": "الحمل",
+	"settings.mediaLoad.utilization": "الاستخدام",
+	"settings.mediaLoad.unknown": "غير معروف",
+	"settings.mediaLoad.notPublished": "لا ينشره برنامج التشغيل هذا",
+	"settings.mediaLoad.sessions": "الجلسات المرتبطة",
+	"settings.mediaLoad.noSessions": "لا توجد جلسات مرتبطة",
+	"settings.mediaLoad.owner": "PID {pid} · الفهرس {index}",
+	"settings.mediaLoad.sample": "العينة",
+	"settings.mediaLoad.stale": "آخر قراءة · لا يتم التحديث",
+	"settings.mediaLoad.legacy": "تتوفر قراءة المُرمِّز القديمة فقط. الاستخدام وملكية الجلسات غير معروفين.",
+	"settings.mediaLoad.accounting": "يحسب الحمل الوقت من التنفيذ إلى جمع النتيجة؛ ويحسب الاستخدام وقت العتاد. قد تتجاوز حسابات الطوابير متعددة الأنوية 100%. هذه القراءات لا تقيس سلامة البث.",
+	"settings.mediaLoad.ownershipNote": "يحدد PID المهمة التي أنشأت الجلسة، وليس بالضرورة العملية. الجلسات مرتبطة بهذا الجهاز؛ ولا يحدد ذلك النواة التي تنفذ العمل أو يميز المعاينة من البرنامج المباشر.",
 	"a11y.close": "إغلاق",
 	"a11y.skipToContent": "تخطّ إلى المحتوى الرئيسي",
 	"advanced.active": "نشط",
