+++ packages/i18n/messages/de.json
@@ -1,5 +1,25 @@
 {
 	"$schema": "https://inlang.com/schema/inlang-message-format",
+	"settings.mediaLoad.title": "Medienlast",
+	"settings.mediaLoad.description": "Treiberwerte je Kern und gebundene Sitzungen",
+	"settings.mediaLoad.details": "Mediendetails",
+	"settings.mediaLoad.coreCount": "Kerne: {count}",
+	"settings.mediaLoad.encode": "Kodierung",
+	"settings.mediaLoad.decode": "Dekodierung",
+	"settings.mediaLoad.jpeg": "JPEG-Dekodierung",
+	"settings.mediaLoad.rga": "RGA · Bildverarbeitung",
+	"settings.mediaLoad.load": "Last",
+	"settings.mediaLoad.utilization": "Auslastung",
+	"settings.mediaLoad.unknown": "Unbekannt",
+	"settings.mediaLoad.notPublished": "Von diesem Treiber nicht veröffentlicht",
+	"settings.mediaLoad.sessions": "Gebundene Sitzungen",
+	"settings.mediaLoad.noSessions": "Keine gebundenen Sitzungen",
+	"settings.mediaLoad.owner": "PID {pid} · Index {index}",
+	"settings.mediaLoad.sample": "Messung",
+	"settings.mediaLoad.stale": "Letzter Wert · keine Aktualisierung",
+	"settings.mediaLoad.legacy": "Nur der ältere Encoderwert ist verfügbar. Auslastung und Sitzungszuordnung sind unbekannt.",
+	"settings.mediaLoad.accounting": "Last zählt die Zeit von der Ausführung bis zur Ergebnisabholung; Auslastung zählt die Hardwarezeit. Mehrkern-Warteschlangen können über 100 % liegen. Diese Werte messen nicht den Zustand des Streams.",
+	"settings.mediaLoad.ownershipNote": "Die PID bezeichnet die erstellende Task, nicht unbedingt den Prozess. Sitzungen sind an dieses Gerät gebunden; daraus geht weder der ausführende Kern hervor noch eine Zuordnung zu Vorschau oder Live-Programm.",
 	"a11y.close": "Schließen",
 	"a11y.skipToContent": "Zum Hauptinhalt springen",
 	"advanced.active": "Aktiv",
