+++ packages/i18n/messages/en.json
@@ -1,5 +1,25 @@
 {
 	"$schema": "https://inlang.com/schema/inlang-message-format",
+	"settings.mediaLoad.title": "Media load",
+	"settings.mediaLoad.description": "Per-core driver readings and bound session ownership",
+	"settings.mediaLoad.details": "Media details",
+	"settings.mediaLoad.coreCount": "Cores: {count}",
+	"settings.mediaLoad.encode": "Encode",
+	"settings.mediaLoad.decode": "Decode",
+	"settings.mediaLoad.jpeg": "JPEG decode",
+	"settings.mediaLoad.rga": "RGA · image processing",
+	"settings.mediaLoad.load": "Load",
+	"settings.mediaLoad.utilization": "Utilization",
+	"settings.mediaLoad.unknown": "Unknown",
+	"settings.mediaLoad.notPublished": "Not published by this driver",
+	"settings.mediaLoad.sessions": "Bound sessions",
+	"settings.mediaLoad.noSessions": "No bound sessions",
+	"settings.mediaLoad.owner": "PID {pid} · index {index}",
+	"settings.mediaLoad.sample": "Sample",
+	"settings.mediaLoad.stale": "Last reading · not updating",
+	"settings.mediaLoad.legacy": "Only the legacy encoder reading is available. Utilization and session ownership are unknown.",
+	"settings.mediaLoad.accounting": "Load counts run-to-reap time; utilization counts hardware time. Multicore queue accounting can exceed 100%. These readings do not measure stream health.",
+	"settings.mediaLoad.ownershipNote": "PID identifies the creating task, not necessarily the process. Sessions are bound to this device; this does not identify the core executing their work or distinguish preview from the live program.",
 	"a11y.close": "Close",
 	"a11y.skipToContent": "Skip to main content",
 	"advanced.active": "Active",
