+++ packages/i18n/messages/es.json
@@ -1,5 +1,25 @@
 {
 	"$schema": "https://inlang.com/schema/inlang-message-format",
+	"settings.mediaLoad.title": "Carga multimedia",
+	"settings.mediaLoad.description": "Lecturas del controlador por núcleo y sesiones vinculadas",
+	"settings.mediaLoad.details": "Detalles multimedia",
+	"settings.mediaLoad.coreCount": "Núcleos: {count}",
+	"settings.mediaLoad.encode": "Codificación",
+	"settings.mediaLoad.decode": "Decodificación",
+	"settings.mediaLoad.jpeg": "Decodificación JPEG",
+	"settings.mediaLoad.rga": "RGA · procesamiento de imagen",
+	"settings.mediaLoad.load": "Carga",
+	"settings.mediaLoad.utilization": "Utilización",
+	"settings.mediaLoad.unknown": "Desconocido",
+	"settings.mediaLoad.notPublished": "Este controlador no lo publica",
+	"settings.mediaLoad.sessions": "Sesiones vinculadas",
+	"settings.mediaLoad.noSessions": "Sin sesiones vinculadas",
+	"settings.mediaLoad.owner": "PID {pid} · índice {index}",
+	"settings.mediaLoad.sample": "Muestra",
+	"settings.mediaLoad.stale": "Última lectura · sin actualizar",
+	"settings.mediaLoad.legacy": "Solo está disponible la lectura antigua del codificador. Se desconocen la utilización y la titularidad de las sesiones.",
+	"settings.mediaLoad.accounting": "La carga cuenta el tiempo desde la ejecución hasta la recogida del resultado; la utilización cuenta el tiempo de hardware. Las colas multinúcleo pueden superar el 100 %. Estas lecturas no miden la salud de la transmisión.",
+	"settings.mediaLoad.ownershipNote": "El PID identifica la tarea creadora, no necesariamente el proceso. Las sesiones están vinculadas a este dispositivo; esto no identifica el núcleo que ejecuta su trabajo ni distingue la vista previa del programa en directo.",
 	"a11y.close": "Cerrar",
 	"a11y.skipToContent": "Saltar al contenido principal",
 	"advanced.active": "Activo",
