+++ packages/i18n/messages/fr.json
@@ -1,5 +1,25 @@
 {
 	"$schema": "https://inlang.com/schema/inlang-message-format",
+	"settings.mediaLoad.title": "Charge multimédia",
+	"settings.mediaLoad.description": "Mesures du pilote par cœur et sessions rattachées",
+	"settings.mediaLoad.details": "Détails multimédias",
+	"settings.mediaLoad.coreCount": "Cœurs : {count}",
+	"settings.mediaLoad.encode": "Encodage",
+	"settings.mediaLoad.decode": "Décodage",
+	"settings.mediaLoad.jpeg": "Décodage JPEG",
+	"settings.mediaLoad.rga": "RGA · traitement d’image",
+	"settings.mediaLoad.load": "Charge",
+	"settings.mediaLoad.utilization": "Utilisation",
+	"settings.mediaLoad.unknown": "Inconnu",
+	"settings.mediaLoad.notPublished": "Non publié par ce pilote",
+	"settings.mediaLoad.sessions": "Sessions rattachées",
+	"settings.mediaLoad.noSessions": "Aucune session rattachée",
+	"settings.mediaLoad.owner": "PID {pid} · indice {index}",
+	"settings.mediaLoad.sample": "Échantillon",
+	"settings.mediaLoad.stale": "Dernière mesure · sans mise à jour",
+	"settings.mediaLoad.legacy": "Seule l’ancienne mesure de l’encodeur est disponible. L’utilisation et l’appartenance des sessions sont inconnues.",
+	"settings.mediaLoad.accounting": "La charge compte le temps entre l’exécution et la récupération du résultat ; l’utilisation compte le temps matériel. Les files multicœurs peuvent dépasser 100 %. Ces mesures n’indiquent pas la santé du flux.",
+	"settings.mediaLoad.ownershipNote": "Le PID désigne la tâche créatrice, pas nécessairement le processus. Les sessions sont rattachées à ce périphérique ; cela n’identifie ni le cœur exécutant leur travail ni leur rôle dans l’aperçu ou le programme en direct.",
 	"a11y.close": "Fermer",
 	"a11y.skipToContent": "Aller au contenu principal",
 	"advanced.active": "Actif",
