+++ packages/i18n/messages/hi.json
@@ -1,5 +1,25 @@
 {
 	"$schema": "https://inlang.com/schema/inlang-message-format",
+	"settings.mediaLoad.title": "मीडिया लोड",
+	"settings.mediaLoad.description": "प्रत्येक कोर की ड्राइवर रीडिंग और उससे जुड़े सत्र",
+	"settings.mediaLoad.details": "मीडिया विवरण",
+	"settings.mediaLoad.coreCount": "कोर: {count}",
+	"settings.mediaLoad.encode": "एन्कोडिंग",
+	"settings.mediaLoad.decode": "डिकोडिंग",
+	"settings.mediaLoad.jpeg": "JPEG डिकोडिंग",
+	"settings.mediaLoad.rga": "RGA · छवि प्रसंस्करण",
+	"settings.mediaLoad.load": "लोड",
+	"settings.mediaLoad.utilization": "उपयोग",
+	"settings.mediaLoad.unknown": "अज्ञात",
+	"settings.mediaLoad.notPublished": "यह ड्राइवर प्रकाशित नहीं करता",
+	"settings.mediaLoad.sessions": "जुड़े सत्र",
+	"settings.mediaLoad.noSessions": "कोई जुड़ा सत्र नहीं",
+	"settings.mediaLoad.owner": "PID {pid} · सूचकांक {index}",
+	"settings.mediaLoad.sample": "नमूना",
+	"settings.mediaLoad.stale": "अंतिम रीडिंग · अपडेट नहीं हो रहा",
+	"settings.mediaLoad.legacy": "केवल एन्कोडर की पुरानी रीडिंग उपलब्ध है। उपयोग और सत्र का स्वामित्व अज्ञात हैं।",
+	"settings.mediaLoad.accounting": "लोड निष्पादन से परिणाम संग्रह तक का समय गिनता है; उपयोग हार्डवेयर का समय गिनता है। मल्टीकोर कतार का लेखा 100% से अधिक हो सकता है। ये रीडिंग स्ट्रीम की स्थिति नहीं मापतीं।",
+	"settings.mediaLoad.ownershipNote": "PID बनाने वाले कार्य को पहचानता है, जरूरी नहीं कि प्रक्रिया को। सत्र इस उपकरण से जुड़े हैं; इससे काम चलाने वाले कोर की पहचान या प्रीव्यू और लाइव कार्यक्रम में अंतर नहीं होता।",
 	"a11y.close": "बंद करें",
 	"a11y.skipToContent": "मुख्य सामग्री पर जाएँ",
 	"advanced.active": "सक्रिय",
