+++ packages/i18n/messages/ja.json
@@ -1,5 +1,25 @@
 {
 	"$schema": "https://inlang.com/schema/inlang-message-format",
+	"settings.mediaLoad.title": "メディア負荷",
+	"settings.mediaLoad.description": "コア別のドライバー測定値と紐づくセッション",
+	"settings.mediaLoad.details": "メディアの詳細",
+	"settings.mediaLoad.coreCount": "コア数: {count}",
+	"settings.mediaLoad.encode": "エンコード",
+	"settings.mediaLoad.decode": "デコード",
+	"settings.mediaLoad.jpeg": "JPEGデコード",
+	"settings.mediaLoad.rga": "RGA · 画像処理",
+	"settings.mediaLoad.load": "負荷",
+	"settings.mediaLoad.utilization": "使用率",
+	"settings.mediaLoad.unknown": "不明",
+	"settings.mediaLoad.notPublished": "このドライバーは公開していません",
+	"settings.mediaLoad.sessions": "紐づくセッション",
+	"settings.mediaLoad.noSessions": "紐づくセッションなし",
+	"settings.mediaLoad.owner": "PID {pid} · インデックス {index}",
+	"settings.mediaLoad.sample": "サンプル",
+	"settings.mediaLoad.stale": "最終測定値 · 更新停止中",
+	"settings.mediaLoad.legacy": "従来のエンコーダー測定値のみ利用できます。使用率とセッションの所有者は不明です。",
+	"settings.mediaLoad.accounting": "負荷は実行から結果回収までの時間、使用率はハードウェア時間を計上します。マルチコアキューの値は100%を超える場合があります。これらは配信の正常性を示す値ではありません。",
+	"settings.mediaLoad.ownershipNote": "PIDは作成元タスクを示し、必ずしもプロセスではありません。セッションはこのデバイスに紐づきますが、処理を実行するコアやプレビューとライブ番組の区別は示しません。",
 	"a11y.close": "閉じる",
 	"a11y.skipToContent": "メインコンテンツへスキップ",
 	"advanced.active": "アクティブ",
