+++ packages/i18n/messages/ko.json
@@ -1,5 +1,25 @@
 {
 	"$schema": "https://inlang.com/schema/inlang-message-format",
+	"settings.mediaLoad.title": "미디어 부하",
+	"settings.mediaLoad.description": "코어별 드라이버 측정값과 연결된 세션 소유 정보",
+	"settings.mediaLoad.details": "미디어 세부 정보",
+	"settings.mediaLoad.coreCount": "코어 수: {count}",
+	"settings.mediaLoad.encode": "인코딩",
+	"settings.mediaLoad.decode": "디코딩",
+	"settings.mediaLoad.jpeg": "JPEG 디코딩",
+	"settings.mediaLoad.rga": "RGA · 이미지 처리",
+	"settings.mediaLoad.load": "부하",
+	"settings.mediaLoad.utilization": "사용률",
+	"settings.mediaLoad.unknown": "알 수 없음",
+	"settings.mediaLoad.notPublished": "이 드라이버에서 제공하지 않음",
+	"settings.mediaLoad.sessions": "연결된 세션",
+	"settings.mediaLoad.noSessions": "연결된 세션 없음",
+	"settings.mediaLoad.owner": "PID {pid} · 인덱스 {index}",
+	"settings.mediaLoad.sample": "샘플",
+	"settings.mediaLoad.stale": "마지막 측정값 · 갱신 중단",
+	"settings.mediaLoad.legacy": "기존 인코더 측정값만 사용할 수 있습니다. 사용률과 세션 소유 정보는 알 수 없습니다.",
+	"settings.mediaLoad.accounting": "부하는 실행부터 결과 수집까지의 시간이고 사용률은 하드웨어 시간입니다. 멀티코어 큐 계산값은 100%를 넘을 수 있습니다. 이 값은 스트림 상태를 측정하지 않습니다.",
+	"settings.mediaLoad.ownershipNote": "PID는 생성한 태스크를 식별하며 반드시 프로세스를 의미하지는 않습니다. 세션은 이 장치에 연결되지만 작업을 실행하는 코어나 미리보기와 라이브 프로그램의 구분을 나타내지는 않습니다.",
 	"a11y.close": "닫기",
 	"a11y.skipToContent": "본문으로 건너뛰기",
 	"advanced.active": "활성",
