+++ packages/i18n/messages/pt-BR.json
@@ -1,5 +1,25 @@
 {
 	"$schema": "https://inlang.com/schema/inlang-message-format",
+	"settings.mediaLoad.title": "Carga de mídia",
+	"settings.mediaLoad.description": "Leituras do driver por núcleo e sessões vinculadas",
+	"settings.mediaLoad.details": "Detalhes de mídia",
+	"settings.mediaLoad.coreCount": "Núcleos: {count}",
+	"settings.mediaLoad.encode": "Codificação",
+	"settings.mediaLoad.decode": "Decodificação",
+	"settings.mediaLoad.jpeg": "Decodificação JPEG",
+	"settings.mediaLoad.rga": "RGA · processamento de imagem",
+	"settings.mediaLoad.load": "Carga",
+	"settings.mediaLoad.utilization": "Utilização",
+	"settings.mediaLoad.unknown": "Desconhecido",
+	"settings.mediaLoad.notPublished": "Não publicado por este driver",
+	"settings.mediaLoad.sessions": "Sessões vinculadas",
+	"settings.mediaLoad.noSessions": "Nenhuma sessão vinculada",
+	"settings.mediaLoad.owner": "PID {pid} · índice {index}",
+	"settings.mediaLoad.sample": "Amostra",
+	"settings.mediaLoad.stale": "Última leitura · sem atualização",
+	"settings.mediaLoad.legacy": "Apenas a leitura legada do codificador está disponível. A utilização e a propriedade das sessões são desconhecidas.",
+	"settings.mediaLoad.accounting": "A carga conta o tempo entre a execução e a coleta do resultado; a utilização conta o tempo de hardware. Filas multinúcleo podem ultrapassar 100%. Estas leituras não medem a saúde da transmissão.",
+	"settings.mediaLoad.ownershipNote": "O PID identifica a tarefa criadora, não necessariamente o processo. As sessões estão vinculadas a este dispositivo; isso não identifica o núcleo que executa o trabalho nem distingue a prévia do programa ao vivo.",
 	"a11y.close": "Fechar",
 	"a11y.skipToContent": "Pular para o conteúdo principal",
 	"advanced.active": "Ativo",
