+++ packages/i18n/messages/zh.json
@@ -1,5 +1,25 @@
 {
 	"$schema": "https://inlang.com/schema/inlang-message-format",
+	"settings.mediaLoad.title": "媒体负载",
+	"settings.mediaLoad.description": "各核心的驱动读数及绑定会话归属",
+	"settings.mediaLoad.details": "媒体详情",
+	"settings.mediaLoad.coreCount": "核心数：{count}",
+	"settings.mediaLoad.encode": "编码",
+	"settings.mediaLoad.decode": "解码",
+	"settings.mediaLoad.jpeg": "JPEG 解码",
+	"settings.mediaLoad.rga": "RGA · 图像处理",
+	"settings.mediaLoad.load": "负载",
+	"settings.mediaLoad.utilization": "利用率",
+	"settings.mediaLoad.unknown": "未知",
+	"settings.mediaLoad.notPublished": "此驱动未发布该数据",
+	"settings.mediaLoad.sessions": "绑定会话",
+	"settings.mediaLoad.noSessions": "无绑定会话",
+	"settings.mediaLoad.owner": "PID {pid} · 索引 {index}",
+	"settings.mediaLoad.sample": "采样",
+	"settings.mediaLoad.stale": "上次读数 · 未更新",
+	"settings.mediaLoad.legacy": "仅有旧版编码器读数。利用率和会话归属未知。",
+	"settings.mediaLoad.accounting": "负载统计从运行到结果回收的时间；利用率统计硬件时间。多核队列统计可能超过 100%。这些读数不衡量直播流的健康状况。",
+	"settings.mediaLoad.ownershipNote": "PID 标识创建会话的任务，不一定是进程。会话绑定到此设备；这并不标识执行工作的核心，也不区分预览和直播节目。",
 	"a11y.close": "关闭",
 	"a11y.skipToContent": "跳转到主要内容",
 	"advanced.active": "活跃",
